test(cli): extract Codex shim harness helpers
